+++ packages/types/src/config.ts
@@ -40,15 +40,18 @@ export interface SlidevConfig {
   /**
    * Where to load monaco types from
    *
+   * - `cdn` - load from CDN with `@typescript/ata`
+   * - `local` - load from local node_modules
+   *
    * @default 'local'
    */
   monacoTypesSource: 'cdn' | 'local' | 'none'
   /**
-   * Additional dependencies to load as monaco types
+   * Additional node packages to load as monaco types
    *
    * @default []
    */
-  monacoTypesAdditionalDeps: string[]
+  monacoTypesAdditionalPackages: string[]
   /**
    * Show a download button in the SPA build,
    * could also be a link to custom pdf
